PT-23:Improve Test structure
